v2.2 Full augmentation test

+++ Main.py
@@ -52,7 +52,7 @@ def main():
     parser.add_argument('--augment_pitch_shift', type=int, default=0.0)  # Pitch shifting, put 1 as ±1
     parser.add_argument('--augment_time_stretch', type=float, default=1.0)  # stretch strength, original speed is 1.0
 
-    parser.add_argument('--augment_loudness_test', type=float, default=0.0)
+    parser.add_argument('--augment_loudness_test', type=int, default=0)
     parser.add_argument('--augment_noise_test', type=float, default=0.0)
     parser.add_argument('--augment_pitch_shift_test', type=int, default=0.0)
     parser.add_argument('--augment_time_stretch_test', type=float, default=1.0)
